requirements.txt
python-coveralls
coverage
paho.mqtt.client
